Consolidate test files under ./test
ANy js test related files are moved under the test fodler.
There's no reason to separate test and hleper directories
