+++ implementation/test/DepositRedemptionTest.js
@@ -1,9 +1,9 @@
-const {deployAndLinkAll} = require("../testHelpers/testDeployer.js")
-const {states, bytes32zero, increaseTime} = require("../testHelpers/utils.js")
+const {deployAndLinkAll} = require("./helpers/testDeployer.js")
+const {states, bytes32zero, increaseTime} = require("./helpers/utils.js")
 const {
   createSnapshot,
   restoreSnapshot,
-} = require("../testHelpers/helpers/snapshot.js")
+} = require("./helpers/snapshot.js")
 const {accounts, web3} = require("@openzeppelin/test-environment")
 const [owner] = accounts
 const {BN, constants, expectRevert} = require("@openzeppelin/test-helpers")
