+++ implementation/test/TBTCSystemTest.js
@@ -1,9 +1,9 @@
-const {deployAndLinkAll} = require("../testHelpers/testDeployer.js")
-const {increaseTime} = require("../testHelpers/utils.js")
+const {deployAndLinkAll} = require("./helpers/testDeployer.js")
+const {increaseTime} = require("./helpers/utils.js")
 const {
   createSnapshot,
   restoreSnapshot,
-} = require("../testHelpers/helpers/snapshot.js")
+} = require("./helpers/snapshot.js")
 const {accounts, contract, web3} = require("@openzeppelin/test-environment")
 const {BN, expectRevert} = require("@openzeppelin/test-helpers")
 const {expect} = require("chai")
